refactor: administration tab with admins section
Rename the owners tab to Administration (a container for game-config sections),
with admin management as its first section. Person=admin, tab=Administration
(en+de). i18n moved to administration.admins.*; OwnersPanel -> AdministrationPanel.
API/keys for ownership are unchanged (still owner).
