+++ src/pages/games/GameDetail/GameViewContent.tsx
@@ -28,7 +28,7 @@ import {
   statusColor,
   translateGameStatus,
 } from '../../../utils/gameStatusHelpers';
-import OwnersPanel from '../panels/OwnersPanel/OwnersPanel';
+import AdministrationPanel from '../panels/AdministrationPanel/AdministrationPanel';
 import RankingsPanel from '../panels/RankingsPanel/RankingsPanel';
 import RoundsPanel from '../panels/RoundsPanel/RoundsPanel';
 import TeamsPanel from '../panels/TeamsPanel/TeamsPanel';
@@ -39,7 +39,7 @@ interface GameViewContentProps {
   game: Game;
 }
 
-const GAME_TYPES = ['teams', 'rounds', 'rankings', 'owners'] as const;
+const GAME_TYPES = ['teams', 'rounds', 'rankings', 'administration'] as const;
 
 type GameTab = (typeof GAME_TYPES)[number];
 
@@ -238,7 +238,9 @@ const GameViewContent = ({ game }: GameViewContentProps) => {
           <Tabs.Tab value="teams">{t('gameDetail:tabs.teams')}</Tabs.Tab>
           <Tabs.Tab value="rounds">{t('gameDetail:tabs.rounds')}</Tabs.Tab>
           <Tabs.Tab value="rankings">{t('gameDetail:tabs.rankings')}</Tabs.Tab>
-          <Tabs.Tab value="owners">{t('gameDetail:tabs.owners')}</Tabs.Tab>
+          <Tabs.Tab value="administration">
+            {t('gameDetail:tabs.administration')}
+          </Tabs.Tab>
         </Tabs.List>
 
         <Tabs.Panel pt="md" value="teams">
@@ -253,8 +255,8 @@ const GameViewContent = ({ game }: GameViewContentProps) => {
           <RankingsPanel game={game} />
         </Tabs.Panel>
 
-        <Tabs.Panel pt="md" value="owners">
-          <OwnersPanel game={game} />
+        <Tabs.Panel pt="md" value="administration">
+          <AdministrationPanel game={game} />
         </Tabs.Panel>
       </Tabs>
     </Stack>
